requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
python-dotenv==1.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
nltk==3.9.1
numpy==1.26.2
